backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
anthropic==0.40.0
aiosqlite==0.20.0
python-dotenv==1.0.1
